Move field deref to top level class
