+++ src/main/java/io/quarkus/gizmo2/InstanceFieldVar.java
@@ -1,9 +1,9 @@
 package io.quarkus.gizmo2;
 
-import io.quarkus.gizmo2.impl.Item;
+import io.quarkus.gizmo2.impl.FieldDeref;
 
 /**
  * An instance field variable.
  */
-public sealed interface InstanceFieldVar extends FieldVar permits Item.FieldDeref {
+public sealed interface InstanceFieldVar extends FieldVar permits FieldDeref {
 }
