+++ src/main/java/io/quarkus/gizmo2/creator/BlockCreator.java
@@ -44,7 +44,7 @@ import io.quarkus.gizmo2.desc.FieldDesc;
 import io.quarkus.gizmo2.desc.MethodDesc;
 import io.quarkus.gizmo2.impl.ArrayDeref;
 import io.quarkus.gizmo2.impl.BlockCreatorImpl;
-import io.quarkus.gizmo2.impl.Item;
+import io.quarkus.gizmo2.impl.FieldDeref;
 import io.quarkus.gizmo2.impl.StaticFieldVarImpl;
 import io.quarkus.gizmo2.impl.Util;
 
@@ -415,7 +415,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofStaticFieldVarHandle(sfv.desc());
             args = List.of(expected, update);
             typeDesc = MethodTypeDesc.of(CD_boolean, var.type(), var.type());
-        } else if (var instanceof Item.FieldDeref fd) {
+        } else if (var instanceof FieldDeref fd) {
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), expected, update);
             typeDesc = MethodTypeDesc.of(CD_boolean, fd.instance().type(), var.type(), var.type());
@@ -499,7 +499,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofStaticFieldVarHandle(sfv.desc());
             args = List.of(expected, update);
             typeDesc = MethodTypeDesc.of(var.type(), var.type(), var.type());
-        } else if (var instanceof Item.FieldDeref fd) {
+        } else if (var instanceof FieldDeref fd) {
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), expected, update);
             typeDesc = MethodTypeDesc.of(var.type(), fd.instance().type(), var.type(), var.type());
@@ -533,7 +533,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofStaticFieldVarHandle(sfv.desc());
             args = List.of(arg);
             typeDesc = MethodTypeDesc.of(var.type(), var.type());
-        } else if (var instanceof Item.FieldDeref fd) {
+        } else if (var instanceof FieldDeref fd) {
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), arg);
             typeDesc = MethodTypeDesc.of(var.type(), fd.instance().type(), var.type());
