+++ src/main/java/io/quarkus/gizmo2/impl/Item.java
@@ -1,20 +1,16 @@
 package io.quarkus.gizmo2.impl;
 
 import static io.smallrye.common.constraint.Assert.checkNotNullParam;
-import static java.lang.constant.ConstantDescs.CD_VarHandle;
 
 import java.lang.constant.ClassDesc;
 import java.lang.constant.ConstantDescs;
-import java.lang.constant.MethodTypeDesc;
 import java.util.function.BiFunction;
 
 import io.github.dmlloyd.classfile.CodeBuilder;
 import io.quarkus.gizmo2.Assignable;
 import io.quarkus.gizmo2.Expr;
 import io.quarkus.gizmo2.InstanceFieldVar;
-import io.quarkus.gizmo2.MemoryOrder;
 import io.quarkus.gizmo2.desc.FieldDesc;
-import io.quarkus.gizmo2.impl.constant.ConstImpl;
 
 public abstract non-sealed class Item implements Expr {
     protected final String creationSite = Util.trackCaller();
@@ -228,101 +224,10 @@ public abstract non-sealed class Item implements Expr {
 
     public InstanceFieldVar field(final FieldDesc desc) {
         checkNotNullParam("desc", desc);
-        return new FieldDeref(desc);
+        return new FieldDeref(this, desc);
     }
 
     Item asBound() {
         return bound() ? this : new BoundItem(this);
-    }
-
-    public final class FieldDeref extends AssignableImpl implements InstanceFieldVar {
-        private final FieldDesc desc;
-        private boolean bound;
-
-        private FieldDeref(final FieldDesc desc) {
-            this.desc = desc;
-        }
-
-        protected Node forEachDependency(final Node node, final BiFunction<Item, Node, Node> op) {
-            return Item.this.process(node.prev(), op);
-        }
-
-        public boolean bound() {
-            return bound;
-        }
-
-        protected void bind() {
-            if (Item.this.bound()) {
-                bound = true;
-            }
-        }
-
-        public FieldDesc desc() {
-            return desc;
-        }
-
-        @Override
-        public ClassDesc type() {
-            return desc.type();
-        }
-
-        public String itemName() {
-            return Item.this.itemName() + "." + desc.name();
-        }
-
-        public Item instance() {
-            return Item.this;
-        }
-
-        Item emitGet(final BlockCreatorImpl block, final MemoryOrder mode) {
-            return switch (mode) {
-                case AsDeclared -> asBound();
-                default -> new FieldGetViaHandle(this, mode);
-            };
-        }
-
-        Item emitSet(final BlockCreatorImpl block, final Item value, final MemoryOrder mode) {
-            return switch (mode) {
-                case AsDeclared -> new Item() {
-                    protected Node forEachDependency(final Node node, final BiFunction<Item, Node, Node> op) {
-                        return Item.this.process(value.process(node.prev(), op), op);
-                    }
-
-                    public void writeCode(final CodeBuilder cb, final BlockCreatorImpl block) {
-                        cb.putfield(owner(), name(), desc().type());
-                    }
-
-                    public String itemName() {
-                        return FieldDeref.this.itemName() + ":set";
-                    }
-                };
-                default -> new Item() {
-                    protected Node forEachDependency(Node node, final BiFunction<Item, Node, Node> op) {
-                        return ConstImpl.ofFieldVarHandle(desc)
-                                .process(FieldDeref.this.process(value.process(node.prev(), op), op), op);
-                    }
-
-                    public void writeCode(final CodeBuilder cb, final BlockCreatorImpl block) {
-                        cb.invokevirtual(CD_VarHandle, switch (mode) {
-                            case Plain -> "set";
-                            case Opaque -> "setOpaque";
-                            case Release -> "setRelease";
-                            case Volatile -> "setVolatile";
-                            default -> throw new IllegalStateException();
-                        }, MethodTypeDesc.of(
-                                desc().type(),
-                                Util.NO_DESCS));
-                    }
-
-                    public String itemName() {
-                        return FieldDeref.this.itemName() + ":set*";
-                    }
-                };
-            };
-        }
-
-        public void writeCode(final CodeBuilder cb, final BlockCreatorImpl block) {
-            cb.getfield(owner(), name(), type());
-        }
     }
 }
